Split Quick Tag editor and restore focus timing
